Bug#36624536 asan-failure on shutdown at bootstrap
routertest_component_... which call bootstrap fail the ASAN checks after
the recent switch protobuf-lite to full protobuf.

Background
==========

Bootstrap loads the plugins which it created a config for to expose its
configuration in the cluster's metadata tables.

It loads the plugins, gets their "expose-config" hooks and collects the
config descriptions.

When the "metadata_cache" plugin is loaded, and later unloaded the
reflection-data of the libprotobuf is lazy-loaded at library start, but
not freed when the libprotobuf is unloaded again. This task normally
handled by the "router_protobuf" plugin which calls
ShutdownProtobufLibrary()" in its "deinit" function.

But at bootstrap, the plugins are only loaded and never initialized, and
with "init()", the "deinit()" function isn't called either.

Change
======

- move the "ShutdownProtobufLibrary()" into a "static Scope_guard" which
  is triggered then the router_protobuf plugin is unloaded

Change-Id: I652ec11ac9ab4d0099c70cb9975a07500981f84c