WL#16107 Hypergraph optimizer core cost model
Post-push fix. Some tests failed with the hypergraph optimizer.

Most of the failures were caused by different ordering of the results
in queries without explicit ordering. Ignore ordering when checking
the results of these queries.

In one case the failure happened because a table scan was chosen over
an index range scan on a very small table, resulting in different
locks being taken. Fixed by forcing the use of an index.

Change-Id: I5fc4acd39b5e014cd82766add088250fec3a1c91