WL#16005: Index-level hints for hypergraph optimizer WL#16107: Hypergraph optimizer core cost model
Post-push fix.

The main.index_merge_innodb_hypergraph test fails because some of the
recorded query plans have changed. It's partly because of patch 6 for
WL#16005, which made it more likely to pick an index scan over a table
scan if their costs were very similar (because the optimizer now sees
the index scan before the table scan, and chooses the first candidate
if their cost estimates are indistinguishable). And it's partly
because of WL#16107, which changed the cost estimates so that what
looked like the best plan before no longer does.

Fixed by rerecording the new plans, and also ignore the cost estimates
in more of the EXPLAIN statements.

Change-Id: I98a9cf64bef5f8c86cf7093ea8cc7be0bb4609d8