WL#16107 Hypergraph optimizer core cost model
Post-push fix. Some tests failed with the hypergraph optimizer.

Fixed by ignoring the ordering of the result in some queries with
--sorted_result. In main.mem_cnt_common, the use of temp-table
aggregation is now forced with SQL_SMALL_RESULT in some queries so
that they behave the same way as before. The test used these queries
to test limits on memory used by temporary tables, so it depended on
their use of temp-table aggregation.

Change-Id: I45cee302e6bcd7bad08e56952dabe9c49ddd6c3f