Bug#36625773: Align implementation of repeated preparation and optimization for statement classes.
The function sp_lex_instr::validate_lex_and_execute_core() was enhanced
in WL#16142 with improved cleanup of allocated memory and improved
evaluation logic. In order to consolidate the logic in the classes
sp_lex_instr and Prepared_statement, similar actions are also taken for
Prepared_statement::execute_loop().

A local variable "error" keeps track of the current execution state,
and another local variable "need_reprepare" keeps track of whether
the statement must be reprepared in the next loop iteration.

After a failed execution, we ensure that both memory on the statement
mem_root as well as the execution mem_root is properly cleaned up.

In addition, both functions are updated so that they have a
"while (true)" loop for execution and a section after the loop that is
executed after the loop. This also means that we no longer need a
scope guard.

Change-Id: I617b3ad4c136f2dfcdca1b9cc52e839b761c94d0