Bug#36482778: DO NOT COPY SECONDARY_LOAD FLAG IN CREATE TABLE LIKE
When a table is created using `create table like` command and the source
table has secondary_engine defined, it copies the secondary_load flag
along with other secondary engine attributes. While the secondary engine
attributes are part of table's definition, secondary load flag is a
runtime state and should not be copied while creating the table.

Change-Id: I79a4e5283f983dc3a922cece628f91842d0c2c1c