WL#16131 : Support foreign key implicit references to primary key
MySQL does not support foreign key specification that implicitly
reference the primary key of the referenced table.

This WL aims to introduce support for implicit foreign keys
in accordance with the SQL standard, thereby enabling the
creation of foreign keys that automatically reference the
primary key of the referenced table.

1. After parsing, if foreign key referenced column list is found
empty, add referenced table primary key columns to the list.
2. Added new test cases to main.foreign_key

Change-Id: I44917315dcd538f6b6fb82a45aa2c22197b6992a