Bug#33503328: Specific table schema causes table to become corrupted during MySQL 8 upgrade
Summary

Upgrading a table

  CREATE TABLE t1 (c1 int GENERATED ALWAYS AS (0),
                   c2 int) ENGINE=MyISAM;

from 5.7 to 8.x would fail. This would only occur for a
mix of regular columns and generated columns and when using
MyISAM.

The reason for this was that during that particular type
of upgrade for use with the data dictionary, we would align
NULL info one way, and the other way everywhere else. This
was caught by our sanity checks.

Analysis

Using our above table as an example: We always use two loops
to create columns; the first loop processes regular columns,
while the the second loop processes generated columns.
Therefore, we'd process c2 (a natural column) before c1 (a
generated column).

In older code, we'd count the NULL position up in order of
column creation so that c2 (which is the second column in the
CREATE statement, but the first column to be processed on
account of being the first regular column) will be processed
first and get the lower NULL position; c1 will be processed
second (on account of being a generated column and getting
created in the second loop) and get the higher NULL position.

In short, we process all regular columns first, then all
generated columns, and count up NULL position the entire
time.

In fill_columns_from_dd(), we processed columns in the same
order (first all natural columns, then all generated columns),
but would increase the NULL position even when no field was
processed (e.g. when skipping a generated column in the first
loop that only processes regular columns). Therefore, the
NULL positions ended up wrong and tripped our sanity check.

This patch aligns the NULL ordering so that the positions
generated in fill_columns_from_dd() are compatible with the
ordering created by earlier servers and indeed with the
behavior elsewhere in contemporary servers.

Change-Id: I7bd18e0822c1d86b9dd322a5d994200078ebf84c