Bug#35019884: set persist_only thread_pool_max_transactions_limit broken
Problem: Not possible to use SET PERSIST with
thread_pool_max_transactions_limit.

Root cause: Values which have been set by SET PERSIST are applied after
the Thread Pool plugin has been initialized in the restart sequence.
This means that the application of persisted values is seen and treated
as an update by the Thread Pool plugin variable code.

There was a special restriction on updates to max_transactions_limit
which ensured that it could only be changed from one non-zero value to
another non-zero value. This because the value 0 has a special meaning
and indicates that the max_transactions_limit feature is turned off.

This restriction was introduced because it was previously required that
the static variable thread_pool_dedicated_listener had been set in order
use the max_transactions_limit feature. So this restriction tried to
make the enable/disable of max_transactions_limit static, while still
letting it be changed dynamically if it was enabled when the server was
started.

After wl#16132 the issues dedicated_listener was meant to address are no
longer present so there is no need to keep this restiction, except that
it should not be possible to disable max_transactions_limit if the
server was started with dedicated_listeners=ON. A new error message is
reported in this case (instead of reporting a waring and retaining the
previous value).

Solution: Remove the restriction on updates of
thread_pool_max_transactions_limit, except when attempting to disable it
when dedicated_listners is ON, and report error in this case.

Change-Id: Ifda365c76ceb257ecccf82a6fd37a82fbc431380