Bug#36423829: Missing time_zone function
A CREATE TABLE statement with a CHECK constraint that involved
a CAST with an AT TIME ZONE clause failed with the confusing error
"Missing time zone function". The reason for this is because the CHECK
constraint text is regenerated from the internal AST and then
misinterpreted, due to a missing print() function for class
Item_func_at_time_zone.

The fix for the problem is to implement this function. This means
that the correct CHECK constraint text is stored in the data dictionary
along with the rest of the CREATE TABLE definition.

Change-Id: If791dfcc959709f85b9b33f1318c940b11a3d10c