WL#15853 Join order hints for hypergraph optimizer [1 of 3]
The STRAIGHT_JOIN hint is supported in hypergraph for the syntax

SELECT ... FROM <tablename> STRAIGHT_JOIN <tablename> ...

If the STRAIGHT_JOIN hint is specified, the non-associative join
type STRAIGHT_INNER_JOIN is used for that join. The STRAIGHT_JOIN
support is now extended to handle the two other syntax options which
can be used to specify straight joins.

The STRAIGHT_JOIN handling is extended to use STRAIGHT_INNER_JOIN for
the case where the straight join is specified using either of these
syntax forms.
1. SELECT STRAIGHT_JOIN ...
2. SELECT ... /*+ JOIN_FIXED_ORDER() */ ...

The parser sets the SELECT_STRAIGHT_JOIN flag if either of these are
used. Support for these forms is added by expanding the check for
STRAIGHT_INNER_JOIN in hypergraph to include a check for the
SELECT_STRAIGHT_JOIN flag.

Change-Id: Ic82e7ead12951caf3057c87b0dcf88d47b2df789