WL#15853 Join order hints for hypergraph optimizer [3 of 3]
Add support for semijoins, antijoins and outer joins in join order
hints.

1. Add conflict handling and merging for multiple join order hints
in the same query, when the hints have common tables.
2. Handle nested tables in semijoins or antijoins by searching all
the nested tables, if any, while matching a table to a hint table.
If the table matches, the top-level table is added in the hinted sort
order. In addition, the nested tables are sorted in place as per the
hints. This is done to ensure that no table moves from one nest to
another as a result of hints. This also ensures that hinted orders do
not override outer joins.
3. If no root candidates are produced due to optimiser hints, remove
hints one by one from end of list and keep retrying until at least
one root candidate is proposed.
4. Add tests for nested tables, outer joins, lateral joins and
multiple hints in the same query, both conflicting and compatible.

Change-Id: Ic7c08b244305ab9c2836d33d324ca17935a915de