Bug#36652044: Invalid usage of log_errlog() macro in service_pfs_notification.cc
Two error logging calls within "storage/perfschema/service_pfs_notification.cc"
are using log_errlog() macro incorrectly.
The problem is that this macro expects one of the logging priority level values
as the 1st parameter, and that is not being the case.

The fix adds ERROR_LEVEL as the 1st parameter, additionally debug-mode assert
has been added to logging API to validate the priority value for each
logging call.

Change-Id: I6553b43f9359071409ab958b7873e9dde89b3800