Bug#36654846 test hangs as signal handler is not started yet
Some tests hangs if run on slow machines as the test-running tries to
terminate the processes with a SIGTERM after some test-condition has
succeeded even though the excutables haven't set the signal handler yet.

That leads to a hang as the signal is ignored and never sent a again.

Change
======

- use spawn_router().wait_for_sync_point(RUNNING) to wait until
  signal handler is installed.

Change-Id: If7ed60988f33cd600b16ec6939985c629bc06325