Bug#36625752 View with ORDER BY misses a row
As pointed out by the bug reporter, premature optimization leads us to
lose information about the ORDER BY, so that when the view information
is reparsed for use in a SELECT from the view, we see only an empty
table and accordingly produce an empty result set. Without the view
present, we see a single result row for an implicitly grouped query -
as expected.

The optimization removed the ORDER BY when resolving the query of the
view, since it is redundant in the presence of a single row result
set. Normally (when no view is involved) we keep the information that
the query is implicitly grouped even though the ORDER BY is been
optimized away. This is information lost for the view, hence the bug.

Solution: defer the optimization when analyzing the view until we
re-resolve it at view use time, i.e. we make sure the ORDER BY in the
view's definition isn't optimized away.

Change-Id: I22fd3d6a27aba9fb413748234f26667d2b1d9f87