Bug#36239362 - WL16065 Error while selecting from LH table with U_S_E=ON
Issue here is, for a statement executed using Regular Statement
Handle interface, request to execute statement from Secondary
engine is not handled and an error is reported to a client.

Currently, the Regular Statement Handle interface doesn't support
handling request execute statement in another storage engine.
This means that for normal SQL statements, SP SQL statement and
prepared statements, if the execution needs to be transferred to
another storage engine, it happens seamlessly. However, with the
Regular Statement Handle, there's no option to retry the statement
using the another storage engine. As a consequence, the user
encounters an error message.

To fix this issue, implemented code to handle request to execute
statement from another storage engine for Regular Statement Handle
interface. The code now checks if the statement can be executed
from another storage engine. If compatible, the statement is
automatically re-executed using the alternative engine.

It is important to note that the Prepared Statement Handle
interface already handles requests to execute statements from
another storage engine.

Change-Id: I8368cdbfe40cfa129227c901dde9ba41e770c263