Bug#36578613: Hypergraph result difference when querying I_S.TABLES
Some query plans combining hash joins and sorting caused wrong results
with the hypergraph optimizer.

The problem occurred when two SORT paths were stacked on top of each
other without any materialization step in between, and the sorting
required row IDs, and the input to the lower SORT path came from a
hash join.

The cause of the problem is that the sorting iterators are not
updating the row ID in handler::ref when they have positioned the
handler on the row to return. This makes the next sort see the same
row ID for every row, and it therefore returns the same row multiple
times instead of sorting them.

Fixed by making FindTablesToGetRowidFor() take into account that row
IDs are not preserved through SORT the same way as they are through
HASH_JOIN, so that any SORT or HASH_JOIN above a SORT will fetch row
IDs afresh from the handler.

Change-Id: I15d83fbfc748d7cf26bcaecb2f826e485b193c39