Bug#36658669 mysql_server_mock sporadically does not shutdown on SIGTERM
The mysql_server_mock sometimes doesn't shutdown when it receives a
SIGTERM.

Stacktraces show that the mock-server waits on:

  server_mock::MySQLServerMock::run () at mock_server/src/mysql_server_mock.cc:318

which is a condition-variable that waits for all connections to be
closed.

Background
==========

The hang appears as:

1. a connection is held open by the client
2. the SIGTERM leads to close_all_connections being called which calls
   cancel() on each connection, aborting any waiting IO, expecting that
   it leads to a connection close on each connection.

If the connection is _not_ waiting on IO, it does NOT notice the request
for shutdown and waits for the client to send something, while the
client waits for a shutdown of the mock-server. A deadlock.

Change
======

- add a is_terminated_(true) to track that a connection shall be closed
- check is_terminated_ before entering an async_read()

Change-Id: Idc0cb32fb5b0d67a42ac30644f4a80abbd4a2117