Bug#36341532 End range check not working with multi-value indexes
Wnen scanning an index using a range condition, the "end range check"
optimization is used to reduce the number of traversed and prefetched
records. To check the range condition when scanning a secondary index
based on a  virtual column in the clustered record, it copies the
materialized value of this column from the secondary index, converting
it to MySQL format. However, this copy operation cannot be performed
on multi-value virtual columns.

To fix the issue, the check is skipped for indexes with multi-value
columns, disabling the optimization.

Change-Id: If5e0097a246667404f361638e3dfb130236d6c9f