Bug#35930969 MDS offloads to HW based on Incorrect Costings                           from Optimizer
During optimization, MySQL chooses an initial plan based on different
parameters like access cost, index usability, coverage, etc. During
this phase ORDER BY and LIMIT are not considered for the plan
creation and access choice. Later during ORDDER BY optimization
test_if_cheaper_ordering() chooses index scan as a replacement
for sort to get the results in order. This change in the plan is not
reflected in the cost data present in the plan. As a result, simple
query with ORDER BY and LIMIT will have the cost of sorting the
entire table to fetch few LIMIT rows. This cost value is used in
comparision with secondary_engine_cost_threshold to decide Heatwave
offloading. This could lead to offloading of the queries even when
they could perform much better on InnoDB engine.

Fix: Update the table access cost if index is choosen in
test_if_cheaper_ordering() to provide order. And update query cost
accordingly for single table queries and its present in secondary engine.

Change-Id: Ica649c9db687ef1162964d865a2b5e4ec4c0d069