Bug#36388816: `m_index_cursor.is_positioned()' in temptable::Handler::position at src/handler.
Filesort object is set up in the plan finalization phase. The tables
to be sorted are collected using CollectTables() and passed to
Filesort. If Sort is on top of a temp table aggregation plan, the
FileSort object gets tables not just from the temp table, but also
from all of the plan's children which the Sort is not supposed to
handle. So while setting up table positions for such tables, there are
assert failures.

CollectTables() gets these irrelevant tables because WalkAccessPaths()
doesn't stop at AccessPath::TEMPTABLE_AGGREGATE when
WalkAccessPathPolicy is STOP_AT_MATERIALIZATION. Fix this.

In the old optimizer, a single table is passed to Filsort so this
problem is not present in the old optimizer.

Change-Id: I3048aba5e062aa2dcfaba6dbcfad6ea8ef3a2cd1