Bug#36625968: Long query execution times due to hypergraph optimizer choosing un-indexed NLJ
Query plans with nested loop joins are risky if the inner side of the
join cannot use indexes to limit the number of rows to read. Even
though they are estimated to be slightly cheaper than the
corresponding hash join in some cases, if the row estimates are a
little off, the actual cost of executing the join can multiply many
times. The performance of hash joins degrades more gracefully when the
estimates are off.

In order to avoid these risky plans, the hypergraph optimizer is
changed so that it proposes nested loop joins only if it can push the
join condition to an index lookup on the right hand side of the join.
If it doesn't have an index to use for the join condition, only the
corresponding hash join is proposed.

Exceptions to this are made for cases in which a hash join cannot be
used (mainly cases involving full-text search, lateral derived tables
or table functions). Also, a nested loop join is allowed if either of
the sides is known to return at most one row, in which case a nested
loop join should be at least as good as a hash join, and there's no
risk of row estimates being far off.

Change-Id: Iab22e296a5042a7ac1cce80bacc265c7a2c0b78e