BUG#36609971: rpl_truncate_relaylog_load
Problem
-------

There are two problems when running the rpl_truncate_relaylog_load
1. Test related: mysqlslap tool did not finish execution of the
   FLUSH LOGS, which causes the test to fail when comparing the
   expected state of the mysqld.1 process
2. Unexpected sequence of binlog events which is reported
   when running position based replication with STA.

Analysis / Root-cause analysis
------------------------------

1. Mysqlslap tool did not finish execution of the
FLUSH LOGS, which causes the test to fail when comparing the
expected state of the mysqld.1 process. We need to count
the number of FLUSH statements and wait for them to complete.

2. In the failing test case, relay log sanitization process
did not run. Since warnings do not affect the correct
result of the relay log recovery, we disable the warning.
This problem needs more analysis in a separate bug report:
Bug#36672631.

Solution
--------

1. When running a dedicated mysqlslap command, we insert dummy
data into table calculating the number of executed FLUSH
statements. Before ending the test, we wait for the table
to contain the number of rows equal to the number of executed
flush statements.

2. Relevant warnings are disabled. Opened Bug#36672631

Change-Id: I68b883fcbe52b865b332c848fa0dd7226d219112