WL#16122 Temp table DISTINCT in the hypergraph optimizer
Use temp table with deduplication to materialize rows when DISTINCT or
GROUP BY is used without aggregation. Use the current infrastructure
for replacing items with temp table fields. Propose Materialize
AccessPath with deduplication wherever we propose a Sort method to
deduplicate rows for DISTINCT or GROUP BY.

Added a flag in MaterializePathParameters to indicate the DISTINCT or
GROUP BY context. Accordingly create_tmp_table() is called with
distinct=true or the group param pointing to join->group_list.

If any of the DISTINCT items or GROUP BY items are *visible* items of
type BIT, force deduplication by hash to avoid an existing work-around
of converting BIT fields to long fields for temp tables. Otherwise
data of BIT type would be returned as LONG type to the user.

Slices are used for GROUP BY for the same reason temp table
aggregation had to use slices in hypergraph optimizer: the current
temp table item replacement infrastructure does not support items
inside a subquery. For DISTINCT slices are not required because
ORDER BY is the only thing applied after DISTINCT, and ORDER BY items
are already materialized in the temp table.

Change-Id: I051d56a4ed824da756df2e65a5d28d7ab9a5affc