Bug#35443375 Explain format=json gives error when optimizer hint hypergraph_optimizer=ON
Problem:
EXPLAIN would not use the correct JSON format when SET_VAR optimizer
hint was used to turn the hypergraph optimizer on or off.

Solution:
Change Explain_format_JSON::is_iterator_based() to check the current
state of the THD instead of a static value.
This is safe because explain_json_format_version cannot be changed with
SET_VAR hints, so is_iterator_based() will not change value during
execution.

Change-Id: I956bf3432646fab4e1c345a090665a60ef78fd97