Bug#36506918: Hypergraph: User variable gets wrong value in case of temp table aggregation
With hypergraph optimizer, when a "var:=<column>" style is used to
update a user variable with the non-aggregate column of a SELECT
involving temp table aggregation, the variable didn't get updated
correctly.

If a join field such as "@var:=expr" points to a temp table field, the
var assignment doesn't happen because there is no re-evaluation of the
materialized field once the temp table is initialized. So, rather than
returning the temp table field, return a new Item_func_set_user_var
item that points to temp table field, so that @var gets updated.

Inherited this logic from the old optimizer.

Change-Id: I6898187a8b651869ffdc7c71475ed7f13d696341