Bug#36574298 Background histogram update warnings flood error log
Several users have experienced the error log being flooded with warnings
from background histogram updates. One reported case from a customer saw
as many as 300 000 messages per hour.

The offending warning concerned the failure to acquire metadata locks on
a table. There is a 10-second cooldown on background histogram updates
for any given table. Therefore the above flooding of the error log
should only occur under special circumstances, for example with DDL
operations on a large number of tables. Nevertheless, flooding the error
log with messages from background histogram updates is not helpful for
the user.

This patch throttles messages written to the error log from the
background histogram update operation. The rate is set to at most one
message per minute. This rate should suffice for the user to identify
potential problems with background histogram updates.

With this patch we downgrade all error events that happen during the
background histogram update from priority level error to priority level
warning. The reasoning is twofold: 1) A failure to update statistics is
generally ignorable, so a warning is more suitable than an error, and 2)
throttling of messages in the error log only works out-of-the-box for
error events that have a priority level lower than error.

Change-Id: I62d6cd7acf2d0851605eec362de72e57356ef832