Bug#36683772: Misleading error message when the hypergraph optimizer is unable to find a plan
If the hypergraph optimizer does not find a valid plan for a query,
and no error was raised during planning to suggest why a plan could
not be found, it raises the following error:

ERROR 3889 (HY000): Secondary engine operation failed. Reason: "All
plans were rejected by the secondary storage engine".

This error is misleading if no secondary storage engine was involved.
This patch changes it to the following error for queries that do not
use a secondary storage engine:

ERROR 6417 (HY000): No query execution plan was found for the statement.

This is not expected to happen, but it could happen due to bugs, so in
debug builds an assertion failure is raised. (An assertion failure was
raised even before this patch.)

Change-Id: I563c3d3052c96ff573d07a083093f10a33982ee2