WL#16005: Index-level hints for hypergraph optimizer
Post-push fix:
ProposeAllFullTextIndexScans() was not setting "found_fulltext" even when
a full text index scan was proposed. Because of which a table scan was
chosen (post the new cost model changes) even when a index hint is used
in the query to force the use of a full text index. Correcting that in
this patch.

Change-Id: Ia212c235fd319cc4ff30f176bc56975e4055f70e