Bug#36679652 New UBSAN errors in 3rd party code with Clang 18
Clang 18 UBSAN is stricter than the sanitizers we have used so far.

RPC function pointers are defined in /usr/include/tirpc/rpc/xdr.h
  typedef bool_t (*xdrproc_t)(XDR *, ...);
with the additional comment
  XXX can't actually prototype it, because some take three args!!!

Clang 18 UBSAN will warn when invoking e.g. the generated function
  bool_t xdr_pax_msg_1_9 (XDR *xdrs, pax_msg *objp)
which has a different function signature.

The solution is to disable UBSAN warnings when invoking generated RPC
functions.

Similarly for the 3rd party library lzma:
lzma/LzmaEnc.c:2408:5: runtime error:
  call to function MatchFinder_Init through pointer to incorrect
  function type 'void (*)(void *)'

This is the actual function signature:
   void MatchFinder_Init(CMatchFinder *p)

The solution is to disable sanitizers for the entire library.

Change-Id: Ic0a24b61b9b560f20521ab285b2450aa71c0d077