Bug#36686387 recent regression: crash in Item::split_sum_func2
A regression from

 Bug#35855294 Inconsistent results of SELECT statement with EXISTS clause

A side effect; the new code was supposed to be executed only in the
context of expressions containing window functions. This was not
always the case, as evidenced by this bug (which has a query without a
window function).

Fix: improve the check to avoid interference.

Also fix a regression for windowing for the case of an expression
containing a grouped aggregate, a window and an IN subquery where the
left clause of the subquery contains the grouped aggregate, but no
window function.

Fix: always defer IN subquery evaluation till after windowing in an
expression containing a window function.

Change-Id: I1a8431b664fee1799732ca4aa1686ed67ff5d722