bug#36694848 Configure ndb binlog cache size
Problem:
The ndbcluster plugin subscribes to all changes that occur in NDB and
writes them epoch by epoch to binary log. Each epoch received from NDB
might consist of a large number of changes which are all written to the
binlog transaction cache before flushing them to the binary log. In
order to properly dimension and configure the system the binlog
transaction cache used by the ndb binlog injector must be possible to
configure separately from sessions that normally write transactions of
smaller size. Currently it is only possible to configure the cache size
for all sessions which leads to improper resource allocation for a
MySQL Server used for writing binary log of changes in NDB.

Solution:
Make it possible to configure the binlog transaction cache size
used for ndb binlog injector with a new system variable:
  --ndb-log-cache-size=<value>
The new variable can be changed at runtime and it's default value
is set to 64MB. All in all the new variable is very much like --binlog-
cache-size, just that it only affects the transaction cache used
by ndb binlog injector when writing binary log of changes in NDB.

Change-Id: Ia6b7ddd238a132fdf74248a5616a06316197ba63