Fix for valgrind reporting conditional jump in metadata_schema_version_is_compatible
This is one of the issues reported in
Bug#36641463 Trying to boostrap with --mrs but no schema gives weird error

mysqlrouter_bootstrap allows b/s against standalone (no innodb cluster md)
server instance. The code would check for metadata (version) and
if it fails it would continue in case of standalone setup.
The problem is that this metadata version variable that is still uninitialized
in such case is later used when the config file is being created.

This patch fixes this issue, skipping the use of metadata version
during the config generation in case of standalone setup.

Change-Id: Ib70c83966bfee87288cd2b8fa3377643ba493d64