Bug#35728261 	Autotest testNodeRestart -n WatchdogSlowShutdown T1 fails occasionally
Context:
Data nodes has a mechanism (implemented in
ErrorReporter.cpp :: prepare_to_crash) to prevent multiple threads to
start processing the crash handling in parallel. That works for ndbmtd,
where we can have multiple threads crashing at the same time, but for
ndbd case the prepare_to_crash function is empty since there is only
one exec thread. But, in addition to exec threads, the watchdog thread
can also start the crash handling, if this thread 'crash' at same time
as the exec thread we can have both threads trying to start the crash
handling in parallel.

Problem:
TestNodeRestart -n WatchdogSlowShutdown fail in .2ndbd due to
a race between the Watchdog thread and the signal exec thread trying
to start a crash handling in parallel.

Solution:
Singlethreaded version of prepare_to_crash changed in order to ensure
that, if both Watchdog and exec thread crash 'at same time', only one
thread (the first one) proceed with the crash handling, the second one
is stopped immediately.

This patch also fixes an issue in the FastScheduler implementation
of traceDumpGetJam() used by the watchdog thread when it is the first
thread to start the crash handling. Function uses a local
EmulatedJamBuffer that is not set in the watchdog thread.

Change-Id: I5c9ffb6d06bf74cd38b65268171ed9c9718265e4