Bug#36694210: Remove const_cast in Opt_hints_qb::sort_tables_in_join_order()
Remove the use of const_cast in
Opt_hints_qb::sort_tables_in_join_order() and
MakeRelationalExpressionFromJoinList() by making
sort_tables_in_join_order() return a pointer to const data, so that
there's no need to cast away the const-ness before returning. The
caller doesn't modify the returned object anyway, so it doesn't have
to be mutable.

Other changes while at it:

- Changed the type of the join_list parameter from pointer to
const-ref to make it clearer that it cannot be nullptr.

- Removed the JOIN parameter, since it was only used for getting the
THD, and added a THD parameter instead.

- Delayed the allocation of new_join_list a bit, so that it is not
allocated if we decide early that sorting is unnecessary.

Change-Id: I031cf60ba626780f6c375b73a07f4daed5dbe5a4