Bug#36600793 multi_factor_authentication.mfa_bugs_debug test failing with ASAN
Description
---------------
Address sanitizer complains that variable buf_flush_list_added is
accessed after being destroyed when an invalid option is specified to
sys var authentication_policy.

Background
----------------
We added  a global unique pointer  buf_flush_list_added that is
allocated through buf_pool_init() method.  The intent is to control the
life cycle of this unique pointer by explicitly resetting it in the
buf_pool_free(). This works in most cases but not in error situations
like the one reported in this bug. We must remember that global
variables have static duration they are destroyed (if not already) when
the program terminates.

When the user specify invalid authentication policy, main thread
terminates, so the global unique pointer buf_flush_list_added is
destroyed too by the main thread without giving a chance for other
threads to do the cleanup. When other threads close, some of them
(log check pointer, page cleaner) try to access the already
deleted variable thus this issue is observed.
One option is to convert this global unique pointer to naked pointer
and manage deleting it explicitly. But then main thread should not
abort abruptly in error case without doing the proper cleanup like it
does at other places in the mysqld_main() method.  The reason it is
important because even if we fix this variable, some other thread might
still complain due to something unintended as it is undefined behavior.
So ASAN may complain in some different scenario. Thus main thread must
abort gracefully in this case too like it does in other situations.

Fix
----
- In error condition rather returning 1, call
  unireg_abort(MYSQLD_ABORT_EXIT) which does the proper cleanups.
- Filed a followup Bug#36696958 to add a scopeguard in this method to
  avoid this kind of bugs in future in this method.

Change-Id: I2772c1b03ba8f2a7505c4635b93b85ded96d4799