Bug#36729777: Remove recursion in FindBestQueryPlanInner()
When the hypergraph optimizer does not find a valid plan for a join,
and there are join order hints in the query, it restarts planning with
all join order hints disabled. It does this by calling
FindBestQueryPlanInner() recursively.

FindBestQueryPlanInner() has another mechanism for restarting the
planning, which doesn't use recursion, and that is to return true in
the "retry" output parameter. Having a single mechanism is simpler, so
this patch changes the join order hint case to use that mechanism.

Since there are now two cases which use the "retry" parameter, the
retry logic in FindBestQueryPlan() is also changed to allow more than
one retry, in case more than one reason to retry should occur in a
query. (This is mostly for future-proofing, since there is currently
no know case where both reasons to retry could occur in the same
query.)

Change-Id: I024ad71440caec363c47ccffe28ddf366e17c500